# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BotTalk API
  description: |
    BotTalk 是一个开源的微信推送服务（类似 Server酱），通过微信 iLink Bot 向用户发送消息推送。

    ## 认证方式

    - **SendKey**：用于推送 API，32 位十六进制字符串，支持路径传参、Query 传参和 Bearer Token 三种方式。
    - **Session Cookie**：用于管理 API，通过登录或扫码注册后自动设置，有效期 30 天。

    ## 响应格式

    推送 API 使用 `code` + `message` 风格；管理 API 使用 `success` 布尔值风格。
  version: 1.0.0
  contact:
    name: BotTalk
    url: https://bot-talk.com
servers:
  - url: https://bot-talk.com
    description: 托管服务
  - url: http://localhost:3000
    description: 本地开发 / 自部署
...
components:
  securitySchemes:
    SendKeyQuery:
      type: apiKey
      in: query
      name: key
      description: 通过 query 参数传递 SendKey
    SendKeyBearer:
      type: http
      scheme: bearer
      description: 通过 Authorization Bearer 头传递 SendKey
    SessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: 登录 session cookie
  schemas:
...
    PushResponse:
      type: object
      properties:
        code:
          type: integer
          description: "错误码：0=成功, 40001=Key无效, 40002=无可用通道, 40003=内容为空, 42901=频率超限, 50001=全部失败"
        message:
          type: string
        data:
          type: object
          nullable: true
          properties:
            results:
              type: array
              items:
                $ref: '#/components/schemas/ChannelResult'
    ChannelResult:
      type: object
      properties:
        channel_id:
          type: integer
        status:
          type: string
          enum: [success, failed, skipped]
        reason:
          type: string
          description: 仅 status=skipped 时出现
        token_invalid:
          type: boolean
          description: 仅 status=failed 时可能出现，为 true 表示通道已被标记为 inactive

    # ── 管理 API 通用响应 ──
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
    UnauthorizedError:
      type: object
      properties:
        error:
          type: string
          example: "Unauthorized. Please log in."
...
    Reminder:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        message:
          type: string
        time:
          type: string
          description: "格式 HH:MM 或 YYYY-MM-DDTHH:MM"
        type:
          type: string
          enum: [once, daily, every2min]
        enabled:
          type: integer
          enum: [0, 1]
        send_count:
          type: integer
        max_count:
          type: integer
          description: 最大发送次数，0 表示无限制
        created_at:
          type: string
          format: date-time
...
paths:
...
  /{sendkey}.send:
    get:
      tags: [Push]
      summary: "推送消息（Server酱兼容格式 - GET）"
      description: 通过 URL 路径中的 SendKey 推送消息，兼容 Server酱格式。
      parameters:
        - name: sendkey
          in: path
          required: true
          schema:
            type: string
          description: 用户的 SendKey
        - name: title
          in: query
          schema:
            type: string
          description: 消息标题（与 desp 不能同时为空）
        - name: desp
          in: query
          schema:
            type: string
          description: 消息内容
        - name: channel
          in: query
          schema:
            type: string
          description: "目标通道：留空/default=默认通道, all=所有通道, 1,3,5=指定ID"
      responses:
        '200':
          description: 推送结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushResponse'
    post:
      tags: [Push]
      summary: "推送消息（Server酱兼容格式 - POST）"
      description: 通过 URL 路径中的 SendKey 推送消息，兼容 Server酱格式。
      parameters:
        - name: sendkey
          in: path
          required: true
          schema:
            type: string
          description: 用户的 SendKey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 消息标题（与 desp/message 不能同时为空）
                desp:
                  type: string
                  description: 消息内容
                message:
                  type: string
                  description: 消息内容（desp 的别名）
                channel:
                  type: string
                  description: "目标通道：留空/default=默认通道, all=所有通道, 1,3,5=指定ID"
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                desp:
                  type: string
                message:
                  type: string
                channel:
                  type: string
      responses:
        '200':
          description: 推送结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushResponse'
  /notify:
    get:
      tags: [Push]
      summary: "推送消息（通用格式 - GET）"
      description: 通过 query 参数传递 SendKey 和消息内容。
      security:
        - SendKeyQuery: []
      parameters:
        - name: key
          in: query
          required: true
          schema:
            type: string
          description: SendKey
        - name: title
          in: query
          schema:
            type: string
          description: 消息标题（与 msg/desp 不能同时为空）
        - name: msg
          in: query
          schema:
            type: string
          description: 消息内容
        - name: desp
          in: query
          schema:
            type: string
          description: 消息内容（msg 的别名）
        - name: channel
          in: query
          schema:
            type: string
          description: "目标通道：留空/default=默认通道, all=所有通道, 1,3,5=指定ID"
      responses:
        '200':
          description: 推送结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushResponse'
    post:
      tags: [Push]
      summary: "推送消息（通用格式 - POST）"
      description: 通过 Bearer Token 或 query 参数传递 SendKey，POST body 传递消息内容。
      security:
        - SendKeyBearer: []
        - SendKeyQuery: []
      parameters:
        - name: key
          in: query
          schema:
            type: string
          description: SendKey（也可通过 Authorization Bearer 头传递）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 消息标题（与 message/desp 不能同时为空）
                message:
                  type: string
                  description: 消息内容
                desp:
                  type: string
                  description: 消息内容（message 的别名）
                channel:
                  type: string
                  description: "目标通道：留空/default=默认通道, all=所有通道, 1,3,5=指定ID"
      responses:
        '200':
          description: 推送结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushResponse'

  # ════════════════════════════════════════════════════════════════════
  #  Auth API
  # ════════════════════════════════════════════════════════════════════
...
  /api/key:
    get:
      tags: [Key]
      summary: 获取 SendKey
      description: 获取当前用户的 SendKey。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: SendKey
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      send_key:
                        type: string
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /api/key/reset:
    post:
      tags: [Key]
      summary: 重置 SendKey
      description: 生成新的 SendKey，旧 Key 立即失效。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 新的 SendKey
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      send_key:
                        type: string
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'

  # ════════════════════════════════════════════════════════════════════
  #  Channels API
  # ════════════════════════════════════════════════════════════════════
...
  /api/reminders:
    get:
      tags: [Reminders]
      summary: 获取提醒列表
      description: 获取当前用户的所有提醒。
      security:
        - SessionCookie: []
      responses:
        '200':
          description: 提醒列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reminder'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
    post:
      tags: [Reminders]
      summary: 创建提醒
      description: 创建新的定时提醒。
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, time]
              properties:
                title:
                  type: string
                  description: 提醒标题
                message:
                  type: string
                  description: 提醒内容
                time:
                  type: string
                  description: "触发时间，格式 HH:MM 或 YYYY-MM-DDTHH:MM"
                type:
                  type: string
                  enum: [once, daily, every2min]
                  default: once
                  description: "类型：once（默认）、daily、every2min（体验用）"
                max_count:
                  type: integer
                  default: 0
                  description: 最大发送次数，0 表示无限制
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  id:
                    type: integer
                    description: 新创建的提醒 ID
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /api/reminders/{id}:
    delete:
      tags: [Reminders]
      summary: 删除提醒
      description: 删除指定提醒（同时删除关联的发送日志）。仅能删除自己的提醒。
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 提醒 ID
      responses:
        '200':
          description: 删除结果
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'
  /api/reminders/{id}/enable:
    patch:
      tags: [Reminders]
      summary: 启用/禁用提醒
      description: 启用或禁用指定提醒。
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: 提醒 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
                  description: true 启用，false 禁用
      responses:
        '200':
          description: 操作成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: 未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'

  # ════════════════════════════════════════════════════════════════════
  #  Admin API
  # ════════════════════════════════════════════════════════════════════
...